spec: use operator-sdk for kubernetes operator
The Operator-SDK provides both the ability for Ansible and Golang
based operators.

It brings the following advantages:
- Leveraging "EnqueueRequestsFromMapFunc" which allows us to listen
  on a certain resource and requesting a reconcile on anothor resource
  programatically (i.e. listen for endpoint changes and update config)
- Faster reconciliation times as controller-runtime is much faster
  at making changes than an Ansible run.
- Simplified access to the Kubernetes API (for pulling lists of
  resources or a specific resource, much easier than using an
  Ansible lookup).
- Strong typed code which means much less changes at having any
  weird issues discovered at runtime.

It also comes with a few disadvantages:
- Python & Ansible are at the base of Zuul, Golang is not which
  might be a bit unfamiliar for maintainers of the project itself.

Overall, the Ansible operator-sdk is great at what it does for a
much more simple deployment, however, given the sheer amount of
resources that we plan to manage overall (and the complexity of
relationships between them), it is probably for the best to have
the power of a full programming language at our hands for this.

Change-Id: I7c3933d884a79b3a229e2510d982498b65c3b1c3