Update patch set 1
Patch Set 1:

I don't personally have time to commit to improving operator-sdk to have it get all the features we need to continue to build this in Ansible.

I would like to drive the operator and get it up and going as quickly as possible.  So, this is the path to quickest success.

If we do go with the "let's fix operator-sdk for Ansible" first, it could significantly delay all of this.  Just my thought.

Having said that, perhaps not everyone feels that it's a priority that this happens quickly and they're okay with leaving it until we have all the necessary (operator-sdk) upstream tools.

Patch-set: 1