Update patch set 1
Patch Set 1: Code-Review+1

I mentioned this in IRC, I don't really have experience with golang so using ansible-operator is a much faster path for me to onboard. It also has the added bonus, as mnaser said, to maybe uplifting the ansible operator community along with us. However, is that really what we in zuul want to do, I don't know. From the sounds of it, no, we'd like the shortest path forward to getting a working zuul-operator.

With my Ansible hat on, I was hoping it would spark more involvement in Ansible community to engage with us here, however given it has been 6 months with not much movement I don't think we should expect that. And given mnaser maybe have time to volunteer to do work here, maybe this is the best path forward. I'm wanting to make some time to work on this, but it isn't part of my day job, it'll be after hours / weekend stuff.

Patch-set: 1