Add report time to item model
This allows reporters to include the enqueue/dequeue time of an
item.  The item model already hase enqueue/dequeue times which we
use when reporting status, however, a reporter runs right before
the item is dequeued.  So we need one more time value which
corresponds to the start of the reporting phase -- thus report_time
in this patch.

Test coverage for this is added in change
Ida0cdef682ca2ce117617eacfb67f371426a3131.

Change-Id: I093626e098b7ce2deea2b0c25265cb48d38712ad