WIP: Add Item.formatStatusUrl
This method can be used by reporters to generate the best "status"
url to report for an item.  It will report the buildset page if
that is available, or the per-change status page if it is not.

Eventually we plan on making the database required, at which point
we can insert buildsets in the db earlier.  Then we can report the
buildset page in all cases.  By beginning to use this method now,
we can seamlessly upgrade reporters in the future.

Change-Id: Ib0c2ca84f6c4d30f233382048c8885fb73edfeec