Support nodes setting 'auto' python path
The nodepool "python-path" config variable makes it's way through the
node arguments and ends up as the "ansible_python_interpreter"
argument for the inventory when running the job.

Notably, python-3 only distributions require this to be set to python3
to avoid what can often be confusing red-herring errors (e.g. things
like dnf packages appearing to be missing on Fedora, for example [1]).

Upstream is aware of this often confusing behaviour and has made an
"ansible_python_interpreter" value of "auto" to, essentially, "do the
right thing" [2] and choose the right python for the target
environment.  This is available in Ansible >=2.8 and will become
default ~2.12.

This allows, and defaults to, an interpreter value of "auto" when
running with Ansible >=2.8.  On the supported prior releases, "auto"
will be translated into "/usr/bin/python2" to maintain backwards
compatability.  Of course a node explicity setting "python-path"
already will override this.

Nodepool is updated to set this by default with
I02a1a618c8806b150049e91b644ec3c0cb826ba4.

Long-term, this should alleviate the need for setting "python-path"
all together for node configurations unless you are actually doing
something interesting like using a virtualenv or alternative
platforms.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1696404
[2] https://docs.ansible.com/ansible/2.8/reference_appendices/interpreter_discovery.html

Change-Id: I2b3bc6d4f873b7d653cfaccd1598464583c561e7