Add autogenerated tag to Gerrit reviews
When reporting a Gerrit review, add a tag to the ReviewInput
structure so that review messages and approvals carry the
a tag with the prefix "autogenerated:".  The UI (especially in
later versions of Gerrit) can support handling automated messages
specially (this is used by the "Only comments" switch in
polygerrit).  This is available as far back as 2.13 (possibly
older, I haven't checked).

Change-Id: I8c693d7bcd38be4ac0301fcc9a3e97748ead6d4d