Update patch set 5
Patch Set 5:

Can you modify pagure to send the oldrev as well?  That's used to determine whether the ref was created or deleted, or if it was updated, to calculate the differences it encompasses.  Basically there are 3 cases:

1) oldrev is null, newrev is a sha: ref was created
2) oldrev is a sha, newrev is a sha: ref was updated
3) oldrev is a sha, newrev is null: ref was deleted

Both gerrit and github provide this information in nearly the same way, so this should be useful for anyone wanting to integrate with pagure.

Patch-set: 5