Add a more conversational overview to README.rst
Just floating this; it struck me that we drop in "project gating" and
reference "jobs" and Ansible versions on the README.rst but don't
really build up in less formal/more conversational tone the concepts
behind what Zuul does.

I know in a way this replicates concepts.rst, although after I wrote
this and re-read that it was interesting that I'd come at it from
basically the opposite diretion; talking about jobs first and
pipelines later, while the concepts documentation does almost exactly
the opposite.  I don't feel like either is more or less correct; I
think people understand things through different lenses so alternating
viewpoints can be helpful in communicating more widely.

README.rst felt a bit more natural for this because I wrote it
thinking how I'd explain it to a colleague; and my colleague are the
type of people for whom their first exposure to software is likely to
be pulling up the gitea page for it :)

Change-Id: Ia8937f3c54873303ca4c7dc643de41b67ce219be