Always use change url from gerrit api data
Whether using the http or ssh apis for change data get the change url
from the returned json data. This avoids problems with missing scheme
info in the web ui which results the url passed being appended to the ui
url as part of the path (because the browser thinks it is a relative
path).

Change-Id: I9e907c79fa531d15c80e593ca7f7cebc5c516165