Decode k8s ServiceAccount bearer token
The serviceaccount bearer token stored in a kubernetes secret is base64 encoded.
Nodepool provides the original token in base64 encoded form and zuul writes it
to ~/.kube/config for kubectl ansible connector. However, kubectl (eg. v1.14.6)
expects the token as base64 decoded otherwise we get the following error and
ansible playbook fails:
"error: You must be logged in to the server (Unauthorized)"

Change-Id: I9fe68393f1272cc2135441d948d579a4ac41c76d