Update patch set 5
Patch Set 5:

I agree that it would probably be better to handle this asynchronously like we do for merging on enqueue.  However, I think we should implement it in the same way.  Let's not make a new threadpool, let's keep the same model of changing a state variable on the QueueItem, putting a job on the merger gearman queue, and then resuming normal processing until it's complete.

That way when we move to HA schedulers, we don't have to design another system to deal with this.

Patch-set: 5