Make reporting asynchronous (alternative)
Reporting is currently done synchronously in the main loop. In a busy
system reporting items can slow down the main loop considerably.
Especially merge operations can stall zuul for multiple seconds. This
gets worse when there are retries involved in the merges which pause
additionally a few seconds.

Changing the reporting to be asynchronous like merge operations makes
it possible to continue in the event loop while zuul waits for the
results of reporting which reduces overall stalls.

Change-Id: I253d398c3921678362e353e273e9346edb4007d5