Make reporting asynchronous
This is a requirement to enable direct-push via GitHub/Gerrit.

Currently, the reporting is done in the scheduler main loop. When
enabling direct-push via GitHub/Gerrit this could lead to blocking
behaviour of the scheduler as those operations might take some time.
Thus, we must move the processing of reporter actions to their own
thread.

Change-Id: I01a8de81a6a270e666009b8be27392bd44ca8c9d