Update patch set 5
Patch Set 5: Code-Review-1

We had a different side effect of the original patch in production. The hostname config setting silently returned the value of the HOSTNAME environment variable instead of falling back to socket.getfqdn value which broke our live log streaming. 

I'd like to first check that this version of the change mitigates that as well.

Patch-set: 5