Fix kubectl multinode access
When we define multiple Pods for the job then the created `.kube/config`
contains two clusters with the same name, which causes an error. Also the
created Pods have different namespaces and contexts which were not
passed as inventory variables so Ansible could not find nodes except for
the first one.

Change-Id: Ic94856d1f5ff94b3813e4850ed4e63a69fb87eca