Update patch set 9
Patch Set 9:

This version should work for both - production and tests. However, there is a drawback that the GerritConnection will still try to submit the change in production use cases which will fail with a 409 when the change is already pushed. Thus, it will try this request two times. Fixing this will make the tests fail, as the FakeGerritConnection and (Fake)GerritWebserver use the submit call as indicator that the change was merged. Maybe a solution similar to GitHub's merge_modes should work, but so far I did not manage to make both - production and test - use cases work.

Patch-set: 9
Reviewer: Gerrit User 27952 <27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0