DNM: Increase mariadb log verbosity
We are seeing slow mariadb startups in some tests with large gaps in
logging. The default log verbosity is 2, increase it to 9 to try and
fill in those large gaps.

Note we've hacked the zuul config to try and expose this issue more
quickly but as a result this change should not be merged as is.

Change-Id: Ib27681c1e5ac25a0a5f75a6b1bfb24029d0be3c8