Pagure: prevent connector data fetching from API when not needed
This patch removes the burden of contacting the Pagure API
to get connector data (project's api token and webhook token),
if not in the local cache, for every API calls.

Calls that do not require authentication to execute were scoped
under the connector fetch process.

One of the side effects is a more complex and heavy startup.
Indeed Zuul needs to read or request creation of connector data
for each configured Pagure projects before calling the branches
endpoint of every projects.

Second side effect is the impossibility to depends-on other
projects if Zuul is not authorized to create the connector data
on them.

This patch fixes this limitation.

Change-Id: I1e146fdfc4f3a80b1b7a626978bc4801f317025a