Pagure - Hook calls sources whitelist system
Pagure hook payload's signature material is by project,
thus we need project admin access to read the signature
token before validating the payload and handling the event.

This is a limitation in usability and especialy in the
third party CI context where usually the CI system should
not require specific authorizations.

This patch enables the use of a simpler payload verification
system by checking the sender ip address. Zuul operators
set the list of whitelisted sources ip then Zuul will
handle the payload if the source ip is part of the
whitelisted sources.

Change-Id: I7400fb8257d1c10704cbfa985c2ffd89fc810e64