Pagure: remove connectors burden and simplify code
This patch removes the use of the connector system.
Indeed I've figured out that user API token can
be set with the needed rights: pull_request_merge,
pull_request_flag, pull_request_comment. In fact,
the default Pagure configuration (to be set by
Pagure operators) does not allow those right. Then
it is just a matter of configuration.

Recently pagure.io and src.fedoraproject.org operators
have allowed those rights for user API token.

Thus, I think, the connectors system (to get project
scoped API token) was a complex workaround to fix a
configuration issue. I don't see any reason for
a Pagure operator to not allow those rights in the
user API token. Then, with this change, it is assumed
that this driver will be used with a Pagure instance
that allow those ACL rights.

In addition, this patch removes the event payload
signature verification system because it requires the
fetch of the signature token from the connector
endpoint and this again is complex and cumbersome.
The new system to verify the source of event relies
on a white list of IP addresses from which an event
will be accepted.

This change has several benefit:
- No need for the zuul account on a Pagure instance
  to be added as admin collaborator on a project. This
  was required to access the connector endpoint before
  this refactor.
- A zuul account with a token (pull_request_comment,
  pull_request_flag) can now, even not added as
  a collaborator on the projet, report jobs results
  as comment and set a CI flag.
- A third party CI can easily report on a Pagure
  project. The project admin only needs to set the
  web hook url.
- Less API call to Pagure
- Simpler driver code base

Change-Id: Ic38a6774dc4cb9798840a2c874e2d6e9ce16a067