Fix updating change with draft change on top
Zuul fails to update change A that has a draft change B on top. This
happens if zuul has no access to draft changes. In this case there is
an exception that prevents enqueing the event for change A. However in
this case the needed_by change could be ignored.

Change-Id: I6634ce4176f24e0ec606714f6500396e83592d91