Update patch set 35
Patch Set 35: Code-Review-1

It looks like we're changing max_validity_time to max_token_age.  Can we support max_validity_time as a deprecated option for at least one release, and then add an "upgrade" release note for this change which tells people to change the name of the option?

Patch-set: 35
Label: Code-Review=-1