Update patch set 15
Patch Set 15: Code-Review-1

My concern on PS1 remains.  I'm not sure that users expect a "--check-config" option to need to talk to a functioning Zuul system with mergers.  This means that one obvious use case -- running this in a Zuul job to validate configuration, is more complicated than one would expect: you would need to start at least one merger in addition to the scheduler, and the merger and the scheduler may need credentials to talk to their sources.  That's the reason we haven't implemented this so far; simply having Zuul itself report the errors is usually better for most changes.

But if there's some other use where standing up a partial Zuul is worthwhile in order to do this, I'd love to hear about it, and how it would be accomplished.

Patch-set: 15
Label: Code-Review=-1