Update patch set 13
Patch Set 13: Code-Review-1

(1 comment)

The database schema is not being changed, so we're just inserting many more rows into the builds table and expecting the user (via web or api) to filter those out, possibly by checking the result ("RETRY") (See also the inline comment about the "RETRY" result).  This is a little different than the parent change to the mqtt reporter, which adds a retry_builds list.  I think it would be preferable to use a consistent approach.

I can think of at least one place in the code where we have assumed that there is only one build of one job stored in the database.  I believe this change will cause problems with that (there is no testing of this situation since it was not anticipated).  Here's the link:

https://opendev.org/zuul/zuul/src/branch/master/zuul/model.py#L2396-L2402

I think that will need additional filtering to get it back to just the final builds of the jobs.

A negative aspect of this approach is that other consumers may run into the same issue.

Another approach without that problem might be to report the previous builds as attributes on the final builds.  So every buildset retains the characteristic that it has at most one build of each job, but that build may have a list of prior builds.  Eg:

{
  name: jobname,
  result: SUCCESS,
  attempt: 3,
  prior_builds: [
    {name: jobname, result: RETRY, attempt: 1, ...}
    {name: jobname, result: RETRY, attempt: 2, ...}
  ]
}

It may also simplify the model methods to actually implement the in-memory storage this way as well.  I could be convinced either way on the question of whether the SQL database should mirror this (with a "prior_builds" table with a FK relationship of prior_builds.parent_build_uuid) or keep it flat as in this patch and reconstruct the data in the API.

Patch-set: 13
Label: Code-Review=-1