Add auto reconfigure option
This allows Zuul to watch the tenant configuration file for any
changes and reconfigure automatically if the file changes.

This also removes the cleanup on newTenantConfig; it's not
necessary because each test gets a new test class instance, and
if we leave it in place, it causes the auto-reconfig to fire
at the end of the tests which is messy.

Also fix a minor error in an elif chain which was causing
spurious log messages.

Co-Authored-By: James E. Blair <jeblair@redhat.com>
Change-Id: Ie3ba59a557f4f12afbd7866174aeaae419b30baa