Update patch set 13
Patch Set 13:

The approach of storing the previous builds in the final build is not so easy to implement. In a retry case, the build is removed and the scheduler simply starts a new build without knowing about the previous runs.
We would need to store the previous builds anyway (e.g. in the buildset, queue-item?) to add them to the new build.

We would go for the first approach and create a new table to store the retry builds.

Patch-set: 13
Reviewer: Gerrit User 28088 <28088@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0