Ensure correct cleanup on repo update and reset
Zuul was not cleaning/pruning stale refs in the correct order which can
be problematic e.g. with branche names containing slashes.

The issue can be reproduced by creating a branch "foobar" on the remote
that is then also cloned to a local repository. Now, when the remote
branch "foobar" is deleted and another branch "foobar/sub" is created, a
fetch without the prune option will lead to the following exception:

    2020-01-08 09:57:00,934 zuul.Repo  ERROR  Retry 1: Fetch /tmp/tmp4g49pxfl/zuul-test/workspace origin None
    Traceback (most recent call last):
      File "/tmp/zuul/zuul/merger/merger.py", line 225, in _git_fetch
        **kwargs)
      File "/tmp/zuul/.tox/py36/lib/python3.6/site-packages/git/cmd.py", line 545, in <lambda>
        return lambda *args, **kwargs: self._call_process(name, *args, **kwargs)
      File "/tmp/zuul/.tox/py36/lib/python3.6/site-packages/git/cmd.py", line 1014, in _call_process
        return self.execute(call, **exec_kwargs)
      File "/tmp/zuul/.tox/py36/lib/python3.6/site-packages/git/cmd.py", line 825, in execute
        raise GitCommandError(command, status, stderr_value, stdout_value)
    git.exc.GitCommandError: Cmd('git') failed due to: exit code(1)
      cmdline: git fetch --tags origin
      stderr: 'error: cannot lock ref 'refs/remotes/origin/foobar/sub': 'refs/remotes/origin/foobar' exists; cannot create 'refs/remotes/origin/foobar/sub'
    From /tmp/tmp4g49pxfl/zuul-test/upstream/org/project1
     ! [new branch]      foobar/sub -> origin/foobar/sub  (unable to update local ref)'

To fix this problem the Repo.update() method will now pass prune=True
and prune_tags=True when fetching from the remote end.

However, this was not enough to correctly reset the repo. Zuul was still
trying to update local heads to match the remote before cleaning up
stale refs.

The fix for this is to reset the working directory to the default remote
branch that 'origin/HEAD' is referencing. This is necessary since the
locally checked out branch might be stale and must be deleted.

Afterwards all stale local heads that no longer exist on the remote end
can be cleaned up. The last step is to update all local heads to match
the remote.

Change-Id: Ic379848f7f14643a33c43248deb1850ce89cb34c