Fix Gerrit config issue in tutorial demo
When running the tutorial demo system, the Gerrit instance is not
configured correctly (e.g. the canonicalWebUrl is not set to
http://localhost:8080/, sendemail is not disabled).

This is because the Gerrit config file overrides specified in the
Docker compose file (using a series of git config calls as part of
the 'command' keyword for the gerrit service) are not applied during
the initialisation of the container.

This change moves these overrides into a separate script, which is
used to replace the default entrypoint for the Gerrit container
image (gerritcodereview/gerrit).

Tested on a Debian system with Docker version 19.03.5

Change-Id: I6b0f6c5e9ea0c33eb589be7ca91c213c8c63299f