Fix Gerrit config issue in tutorial demo
When running the tutorial demo system, the Gerrit instance is not
configured correctly (e.g. the canonicalWebUrl is not set to
http://localhost:8080/, sendemail is not disabled).

This is because the Gerrit config file changes specified in the
Docker compose file, using a series of git config calls as part of
the 'command' keyword for the gerrit service, are not applied during
the initialisation of the container. This appears to be because the
container's entrypoint is the call to '/var/gerrit/bin/gerrit.sh run'

This change overrides the default entrypoint to resolve this.

Tested on a Debian system with Docker version 19.03.5

Change-Id: I6b0f6c5e9ea0c33eb589be7ca91c213c8c63299f