Support dynamic badges
Zuul currently has a zuul/gated badge that can be linked e.g. in a
readme of a project. This is sufficient for many use cases. However if
a project has periodic jobs that do extended testing which is not
possible in check/gate this is not sufficient. For use cases like
those we can add support for dynamic badges in zuul itself.

This can be done by enhancing the buildsets endpoint to return a badge
for the first element if an image is requested via the accept header.

Change-Id: I449fa9f38ca251522789b6075fbc876d21bd0200