Allow Passing of Jitter Values in TimerDriver
For periodic pipelines to not all trigger at the same moment and thus
potentially bear a lot of load on a system, we could spread the trigger
times by a jitter value. The APScheduler library Zuul uses for timer
triggers supports such jitter values.
This change allows one to pass that jitter value within the cron-like
'time' parameter of the timer trigger.
The optional Jitter parameter can be passed as the last parameter after
the optional 'seconds'.

This change also introduces error-handling in case of invalid numeric
values passed to APScheduler's CronTrigger, which fires an unhandeled
ValueError in such a case. Triggers with invalid values will get
ignored, just like tiggers with an invalid 'cron string' (see line 65 of
the same file).

Change-Id: I4fd5d80c0fb3a4fe27a1687c607774eeed3f9db6